Support a Python 3.10 build in 5.15 without limited API
The original patch worked with limited API, only.

This one adds the missing _Py_Mangle function as in
PySide 6.X and modernizes and fixes the modern embedding
support.

A remaining problem is building with debug Python 3.10 .
There are many refcount errors which need to be found
again, because they all were solved in PySide 6.X .

With the same limited quality of the solution for 5.15 and
the limited API, this is a fix, the debug problem comes later.

Task-number: PYSIDE-1749
Fixes: PYSIDE-1775
Change-Id: I18a9b6756e49850c5d2d2e65b65e1ae934154905
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>