shiboken6: Refactor type system parsing
Instead of using a static variable in
TypeDatabase::parseFile(QIODevice *) to count the invocation depth,
split the parse functions into top level functions taking a file name
or a QIODevice * (for testing) and nested level functions taking an
opaque struct TypeDatabaseParserContext.
The adding of the built-in types happens at the end of the top level
function.

Move the header typesystemparser.h to typesystemparser_p.h to indicate
it is a private class of ApiExtractor and remove a duplicated message.

Pick-to: 6.3
Task-number: PYSIDE-454
Change-Id: I96cfd880cb9b6588e6d55af32ccd6dbf9bebc5b9
Reviewed-by: Christian Tismer <tismer@stackless.com>