setup: fix PySide6.__all__ after the wheel split, augmented
The original fix had a criterion to use the unchanged
module list when a build is active. The determination if a
build is in progress was too setup.py specific.

Using a compiled-in variable solved in a general way.

This patch can be removed completely when the wheel
split method is getting aware of __all__ .

Thanks to Antonio Rojas for the nice hint.

Change-Id: Iaf4cb28056b7f80f2f0c9fb5604c3033bfc1aba5
Fixes: PYSIDE-1890
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 703d975f16aff95bc9014a2689a3ae824b5a552f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>