shiboken6: Move ownership of the metabuilder classes into ApiExtractorResult
ApiExtractorResult kept flat pointer lists of the metabuilder classes
which were owned by AbstractMetaBuilder. This meant
ApiExtractor/AbstractMetaBuilder had to be kept around while the
generators were run.

After some failed attempts to use a QSharedPointer for the metabuilder
classes to fix the ownership, introduce a QExplicitlySharedDataPointer
keeping a class ApiExtractorResultData which maintains ownership of
the classes. ApiExtractorResult can then be used independent of
ApiExtractor/AbstractMetaBuilder.

Pick-to: 6.3
Change-Id: I6f9326d6349ac5c32fd1c7dc8103e961bc353898
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>