Bump version number for dev
The shiboken/pyside version numbers are still at 6.3.0a1
in the development branch. We need to update this to 6.4.0a1
because the PyEnum feature depends on the version number.

For all 6.3 versions, the default setting for new enums is
set to False, which keeps complete compatibility.

For versions beginning with 6.4, the new enums are enabled
by default. Therefore it makes sense to update versions now.

Task-number: PYSIDE-1735
Change-Id: I0a0fac178203fb9ab4790ff511564abc4e191c01
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>