Fix overriding QGraphicsItem::itemChange()
9ef3c0041c489bd9f939c45a571026c3e2227685 introduced meta
type registration for pointers to value classes for signals and slots.

Registering QGraphicsItemGroup* breaks QGraphicsItem::itemChange()
which returns a QGraphicsItem* in ParentChange. When a
QVariant<QGraphicsItemGroup*> is returned, Qt cannot recognize this.

As a quick fix, disable for meta type registration for polymorphic
classes. This needs to be made configureable.

Fixes: PYSIDE-1887
Pick-to: 6.2 6.3
Change-Id: I714236031c2011790cbba58da40af35d05d09d33
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>