shiboken6: Generate includes in deterministic order
The meta builder loops over all functions of a class and adds the
includes required for the arguments to the extra includes of the class
type entry. This apparently had a non-deterministic order.

The extra includes are written by the header generator, but are not
strictly required there.

To fix this, add a separate list TypeEntry::argumentIncludes()
for those argument includes which is written by the CppGenerator.

Introduce a struct IncludeGroup for a list of includes with a title
and stream operator writing them out in sorted order and pass this to
CppGenerator::generateIncludes().

Move the C++ includes to the end.

Remove some spaces.

Pick-to: 6.3
Change-Id: Iea3fdbbc9ace9de7506a58c2f73e5446aa7d42d8
Reviewed-by: Christian Tismer <tismer@stackless.com>