Python 3.10: Fix GIL state crash when collecting QObjects with signals
SignalManager::retrieveMetaObject() calls SbkObject_GetDict_NoRef() to
obtain the ob_dict. This function creates the dict if it does not
exist, causing a GIL violation in some situations (when calling from
some ~QObject from Qt). Acquire the GIL for the creation.

A proper fix might be to add a function that does not enforce
the creation of the dict from SignalManager::retrieveMetaObject().

Pick-to: 6.3 6.2
Fixes: PYSIDE-1919
Change-Id: I3f5f37c4ca92b368763c0dc2a43f16efa004de06
Reviewed-by: Christian Tismer <tismer@stackless.com>