shiboken6: Introduce an attribute for parent ownership heuristics
The parent ownership heuristics cause some unintended ownership
transfers to be generated that can cause crashes. However, completely
removing them would mean to have to explicitly specify it in each
QObject-derived class constructor.

To improve this, introduce an attribute to restrict the heuristics to
classes to which this is applicable.

For compatibility, the attribute is only honored in shiboken 6 if some
class in the type system explicitly specifies it. For shiboken 7,
it will be mandatory.

[ChangeLog][shiboken6] An type system attribute for restricting
the parent ownership heuristics to relevant classes has been
introduced.

The workaround 0e62c4db50472f8b581022c86a6be651158d0cd0
can then be reverted.

Task-number: PYSIDE-1939
Change-Id: I0fb01e9842aecbccaa686ce71b2163b20a97cbf7
Reviewed-by: Christian Tismer <tismer@stackless.com>