PyEnum: Fix the old Enums to produce correct .pyi files, again
Before starting the PyEnum move, I always used Python to check
the validity of everything in a .pyi file. I tried that with
Qt enums as well, but it crashed when tested. Therefore,
I did the transition to Python enums, prematurely.

Later on, it became clear that .pyi files are no longer compatible
to be run in Python at all, and Mypy is the checker, instead.

After removing the Python self-test, the original, correct
enum handling can be restored. Depending on PySide version,
the .pyi files are now correctly created for old or new Enums.

[ChangeLog][PySide6] Enums are now correctly handled in Python
interface files, depending from the version selection.

Change-Id: I79acc4d12b3c16e2624fbb0a6147bbeb40075ebb
Fixes: PYSIDE-1731
Pick-to: 6.3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>