flake8: fix style issues to build_scripts, amended: fix main.py
This cosmetic change caused plain Python builds without virtual env
to break, again. The change was sorting the imports of main.py
in some arbitrary way that caused problems.

It would be much more convenient if changes to the setup scripts
were tested with and without venv or virtualenv. These repeated
errors are an annoying waste of time.

Change-Id: I84335be874cc96128fa192a288a8a7909af13e99
Pick-to: 6.3
Fixes: PYSIDE-1760
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>