PyEnum: Use embedding of the Enum module for better execution speed
The Enum module was optimized several times and is now in good
shape. Old Python versions show massive slowdowns.
Also, Python 3.6 does not support the functional API.

As a general solution, we embed the current enum version to support older
Python versions with the same module. Newer Python versions do a normal import.

As a side effect, we have no longer to treat the old version
Python 3.6 in any special way.

This change is further necessary to be able to remove the old enum version,
since Python 3.6 still exists in CI.

NOTE: Python 3.6 and 3.7 had a hard-to-find refcount error
when creating new enums.

[ChangeLog][shiboken6] Python Enums use the newest implementation for
Python (3.10) for compatibility and speed.

Task-number: PYSIDE-1735
Change-Id: I1f105a6b0fb48e9364cb0654ccd9118d39f68542
Pick-to: 6.3
Reviewed-by: Christian Tismer <tismer@stackless.com>