pyside: Modify QAbstractListModel.data(..) to consider enums
- The function under consideration returns a QVariant, which is used
internally within the C++ code. Sometimes, the value stored in this
QVariant can also be a Python enum/Shiboken enum. When such a case
occurs, the returned QVariant to the C++ side is a QVariant of Python
/Shiboken enum, which cannot be processed from C++ code.

- This is a regression from 108d82d2ca5323652a50706b37f829b9f0457367
which needs some special handling.

- The solution is the convert the python enum (Shiboken enum) to the
corresponding integer type, which is stored into the QVariant and
returned to the C++ side.

Pick-to: 6.3
Fixes: PYSIDE-1930
Change-Id: Id4e9a1caa5a318e2410e584502c6dc6e35e3a657
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>