QEnum: Adapt to the new enum implementation
QEnum was implemented with Python enums, as opposed to the
old C++ enums. Now that support for Python enums is built
into PySide, it is necessary to keep the implementing
Python modules always the same.

This patch prepares QEnum and PyEnum compatibility for the
upcoming embedding of old enum implementations.

As a side effect: Python 3.6 and 3.7 had a hard-to-find
refcount error when creating new enums.

Task-number: PYSIDE-1735
Change-Id: Ic44459c8e319d6d0308c7366beafa6fe7df41b79
Pick-to: 6.3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>