snippet_translate double colon improvements
- Earlier, double colons were converted to dot operator only when the
statement had a QObject class or namespace. For cases with a normal C++
namespace like MyClass::x, it was still translated without modifications
to Python.
- This patch adds an extra statement at the end of snippet_translate(x)
to convert all the remaining scope resolution to dot operator

On top of the above changes, it also addresses a FIXME to handle C++
iterator declaration in Python

Task-number: PYSIDE-1972
Pick-to: 6.3
Change-Id: I45d12954835aaa569d1a4ef15badb366eaff0fe7
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>