PyEnum: Increase compatibility by allowing defaults and old flag names
This patch supports to write

    Qt.Alignment()

instead of

    Qt.AlignmentFlag(0)

Also supported is

    Qt.AlignmentFlag()

which is mapped to

    Qt.AlignmentFlag(0)

This trickery was quite involved since the Python opcodes
needed to be analyzed if we have a parameterless call.
Only in that case, we insert a partial object which supplies the
missing value=0 default.

Changing the implementation of PyEnum was not desired because this
is highly complicated, not portable and even not possible.

The change has been tested with Python 3.6 to 3.11.0b3 .

[ChangeLog][shiboken6] The new Python enums are made as compatible
to the old ones as possible. It is again allowed to use Qt.Alignment()
instead of Qt.AlignmentFlag(0), and a default of 0 is always allowed.

Change-Id: If6a93f8210ff6cae4e38251420e1ad5fffbe42cb
Pick-to: 6.3
Task-number: PYSIDE-1735
Reviewed-by: Christian Tismer <tismer@stackless.com>