shiboken6: Refactor argument conversion for virtual method overrides
Virtual method overrides use Py_BuildValue() to create the argument
tuple for the call with its built-in conversions.

Fold the code for creating the argument and its format string
into a new helper CppGenerator::virtualMethodNativeArg().
Remove ShibokenGenerator::getFormatUnitString().

Change the logic to check for conversion rules from modifications and
convertable primitives first and fall back to shiboken conversions for
the rest, removing the complicated condition duplicated in
ShibokenGenerator::getFormatUnitString() and
CppGenerator::writeVirtualMethodNativeArgs().

This fixes a bug where no conversion was created when passing
smart pointer by values in a virtual functions.

Task-number: PYSIDE-454
Change-Id: I20ef047c89a0d0047f5234a90aae548ca8e6b932
Reviewed-by: Christian Tismer <tismer@stackless.com>