Fix Documentation - Cross Compilation
- Fix symlinks appears as if it were only a part of Option B of
prerequisities. This is now fixed by adding a new sub-heading.
- Added a note about having same version for host and target Qt.
Having 6.3.0 on the host and 6.3.1 on the target, does not work for
cross compilation. The documentation mentions same minor version works
for cross compilation. However, this works for 6.2.4 on host and 6.2.1
on target. Hence, not completely removing the statement.

Change-Id: Ic678dbf094998108c55d916074753db41e9a2e0f
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit a372285581adc26730b7919d878bbe9eb2f64b2e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>