Shiboken: use the Python 3.11 buffer interface for Limited API
Because the stable API includes now the bufferprocs,
we add a warning to remove the special handling when
the lowest version is 3.11 .

Unfortunately, I see no other way to use the new buffer
interface, because things must work with every Python runtime.
But that also does no harm, since this is now in the stable API.

And of course, we can remove the boring version check :)

Also, an old shiboken error workaround could be removed.
XXX No, the shiboken error still exists in RHEL

Change-Id: I2fae8cabb2cf116a7365b9cf45618da5238c2ea7
Pick-to: 6.3
Task-number: PYSIDE-1960
Reviewed-by: Christian Tismer <tismer@stackless.com>