build: Fix a weird build problem on macOS and other platforms
There was a long existing bug in the PySide build system that
had a weird effect on macOS:

* building PySide with Limited API and testing worked fine.
* building PySide without Limited API worked fine too, but gave
  crashes when testing pysidetest. It appears that macos
  duplicates a static Python build, one of them uninitialized.

After a month with no solution, a Linux build had problems, too,
but complained already in the build.
It was solved by passing `--enable-shared` to the Python build.

This `--enable-shared` solved the macOS problem as well.
Further analysis by Alexandru led to the suggestion that the
Python libraries in pysidetest are not needed at all, and by
removing them, the whole problem went away.

This is now confirmed on other platforms as well. The option
`--enable-shared` is no longer needed in any configuration.

The problem was caused by 22e359bc19daf19033332aa2d7bb9726c7f4be14
from 2020-02-18.

Task-number: PYSIDE-1735
Change-Id: I08b8628e75ef9c664be30200ae295cd2e6004370
Pick-to: 6.3
Reviewed-by: Christian Tismer <tismer@stackless.com>