PyEnum: Fix QKeySequence to stay correct when switching IntEnum to Enum
Most PyEnum types will no longer be compatible with int, when we change
from IntEnum to Enum.
This has a number of consequences and leads to improvements
in type representations.

The repr of QKeySequence was actually wrong and misleading,
see the test qkeysequence_test.py .

We change the repr and the indexing to return QKeyCombination like
our competitor does since a long time.

[ChangeLog][PySide6] QKeySequence returns now QKeyCombination instances
when indexed and the new PyEnums are used.

Change-Id: Idaaeb44ca3b6103b3e0416eaeb691ad954745f73
Pick-to: 6.3
Task-number: PYSIDE-1735
Fixes: PYSIDE-2021
Reviewed-by: Christian Tismer <tismer@stackless.com>