PyEnum: Enable the type definition to override all enum types
The new enum support in XML knew IntEnum and IntFlag, only.
After we recognized that there exist three cases that should be
IntFlag, although there is no according "flags" C++ definition
and adding this definition crashes, it became clear that we
must allow to override IntEnum by IntFlag, just for Python.

But the existence of such cases may be not restricted to IntEnum,
it can also occur with normal enums.

This patch prepares such an override. It must be still checked
if there are such cases at all.

Task-number: PYSIDE-1735
Change-Id: I4af1c3153c84f88fbef6ac36e421c47dfb5429a9
Pick-to: 6.3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>