signature: Update, cleanup and install signature test
The signature tests were disabled for a long time.
Meanwhile, a number of unresolved values have piled up.

The reason was partially a weird mixture of intents that
made it unpractical to run the full tests on every platform.
Mamely, macOS was quite unclear about which signatures
have to exist and which don't.

This overhaul does the following:

* The various Shiboken modules were removed from the tested
  signatures

* A pure test is added that only checks for no offending lines

As effect, signatures are tested unconditionally without any
other conditions. The other tests remain disabled.

Task-number: PYSIDE-510
Change-Id: Id9e65ed7324a795966f460e13dd95fd5cebcca6a
Pick-to: 6.3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>