PyEnum: Accelerate item access
With the transition to Python 3.11, access to enum items has
become slower by the replacement of direct item access by
Python properties, involving unnecessarily the use of Python
code.

When we are in charge of Enum item access, we circumvent the
slower property access by using the internal mapping directly.

Task-number: PYSIDE-1735
Change-Id: Iabe045be09df847d9587e9d3f6913e9610f5695e
Pick-to: 6.3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>