Fix random values in QFormLayout.getItemPosition()
In QFormLayout::getItemPosition() on the C++ side, *rolePtr is not set
if row == -1, in which case on the Python side this gets converted to a
random value outside the enum range. Fix this by setting *rolePtr to a
default value in the glue code.

Change-Id: I0c67c771ed017f14b211fa6ccf9c89706152d214
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 5e4a830f2aa5e27e203fc0a4fc482b339d9106a4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>