PyEnum: Finish the flexible Enum/Flag assignment and correct
The former fixed coupling between flags and the Enum/Flag decision
is no longer correct, because there are rare cases which deviate
heavily from the default.

One example is Qt.Modifier: This class is equivalent to
Qt.KeyboardModifier, which is a Python Flag type.

This patch implements the full flexibility of redefining enum
and flag and corrects the Qt.Modifier case.

Task-number: PYSIDE-1735
Change-Id: I7947e722f72b4d7b821bce57586e1793a1089606
Pick-to: 6.3
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>