PyEnum: Simplify code generator because of full forgiveness
PyEnum needed extra signature handling in the parser.py file
and the abstractmetatype.cpp file because we needed to decide
at runtime what names are to be generated.

With the full forgiveness mode, all special handling is
still implemented, but the old versions of the enum
representation can also always be used.

The unexpected funny side effect is that we now can remove
all special handling that has been implemented for these cases.
The way how things are written is no longer relevant, although
the new version is still the only one advertized.

Change-Id: I76efc8ffc646d3a04d974d6776a1c9c2b5ffec70
Pick-to: 6.3
Task-number: PYSIDE-1735
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>