shiboken6: Refactor Handling of CustomConversion
CustomConversion can appear in PrimitiveTypeEntry, ContainerTypeEntry
and ValueTypeEntry. Move the field from the base class TypeEntry
there.

The deprecated QString targetConversionRule() was only implemented for
ValueTypeEntry; move it from the base class TypeEntry
there.

In the original code, CustomConversion was stored as a raw pointer in
TypeEntry. This is bad since TypeEntry are cloneable. Use a
QSharedPointer to prevent crashes.

Change-Id: Ia74219671bbd5792398f9711b4a020f5c9825b1b
Reviewed-by: Christian Tismer <tismer@stackless.com>