Generate documentation for all modules independently
Previously, only the modules that were built in PySide were considered
for documentation, which required one to create a full build just to
build the documentation. One reason for this was that the inheritance
diagram creation determined the base classes by loading the
classes/modules and do introspection. This has been replaced by a
mechanism were shiboken creates a JSON inheritance file.

It is then sufficient to have the modules built in Qt so that the
includes exist for shiboken and qdoc finds the source.

Add a script that retrieves the Qt modules built by looking at the Qt
include path, rearranges them by dependency order and use that as an
input to the documentation checks.

With that, a minimal build of PySide6 suffices to generate the complete
documentation.

The script can be used standalone with an option to print warnings
about modules not built in Qt.

Task-number: PYSIDE-1106
Change-Id: Iade4c6f1d3d255f7052be94d0a6fd647f5ee4ed3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>