Documentation/patch_qhp.py: Strip the module from the index
This makes the index actually useable since it is then
possible to find a class by starting to type the name.

The downside is that there is a conflict with the Qt names
if the C++ documentation is installed, too.

Change-Id: I231a90025b2777fe8ff23fcc9dd669722dcd922e
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit fa38984e53a7a325fd317822f75551379763b810)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>