shiboken6: Fix smart pointers of type <const Pointee>
Use the right function to strip the qualifications from
the type since  <const Pointee> and  <Pointee> are treated identically.

Fixes a regression introduced by change
a262e9bae5dbdef92d5caa0e058a1ad07fa974d3.

This is in principle tested in the smart binding test, but
occurs depending on the order the types are seen, so, add another
test.

Fixes: PYSIDE-2071
Pick-to: 6.3
Change-Id: I838b1ae1dd607095b41018c973093a380f51ab6b
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>