Finally implement the __slots__ feature in PySide
The __slots__ feature never worked for PySide at all.
It was forgotten that __slots__ are implemented as descriptors
which are living at the end of the type instance.

This was solved for normal SbkObjects and for QApplication
as well. Some slight rearrangement of type casts was necessary.

[ChangeLog][PySide6] The long missing support for __slots__
was finally added.

Change-Id: Ieddb92104ab10e80e564b38e053f280f90d7a6a7
Fixes: PYSIDE-1970
Pick-to: 6.3 6.4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>