build: introduce log level
Removed the "quiet" and "verbose" older approach,
and introduced a 3-level configuration so we can switch
between them depending of the amount of output we need.

By default, the log level is set to INFO,
but it can be changed to VERBOSE and QUIET.

The older "--verbose-build" and "--quiet" options were deprecated.

The new way for setting the log level is:

    --log-level=quiet
    --log-level=verbose
    --log-level=info (default)

The default option was made less verbose, controlling the output
of the logging module, and cmake:
for verbose it uses DEBUG, quiet uses ERROR, and no option uses INFO.

Change-Id: Ida0289e3ea0ed09b7489a7513254137bba186fd0
Reviewed-by: Christian Tismer <tismer@stackless.com>