_PepUnicode_AsString: Fix a broken promise
This function was optimized in an incompatible way.
It does not work when the string argument is short-lived.

But: It was found that the equivalent function without Limited API
does not survive short-lived arguments, either. So the broken promise
was still there, but in a different way than assumed ;-)

Fortunately, the replacement function PyUnicode_AsUTF8 is part
of the Limited API since Python 3.10 and we can avoid this hack.
Unfortunately, there is no way to know which runtime-version
this will be, and we must wait until Python 3.10 becomes the
minimum version :(

Change-Id: Id2b1ea3212984a69bef8d71a578825978c59947e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit a8004df5a920e02a783d896af33bb83619492b02)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>