doc: order the sidebar content
As a quick-access to all the documentation,
the sidebar of the documentation was a mix of topics without
any logical order.

Creating directories with an index.rst file,
and putting the content on the right topic toctree
allow us to have a more clear and simple general toc.

Change-Id: I43af890ce988946ababcd575d431fc66704c3e85
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 782acff166d4a2b63846c7ba5493f230cccbd4fc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>