shiboken6: Generate class includes into the wrapper header if needed
When a native wrapper is generated with the protected hack
disabled, the argument includes need to go into the wrapper header
in case inline functions are generated. Introduce a helper function
to ShibokenGenerator to collect all required includes (including
the previously missing conversion operators) and use it for
the header and source generators.

Pick-to: 6.4
Task-number: PYSIDE-2099
Change-Id: If58cd91e3a83d2109207e431f06d6fc1fd04c443
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>